+++ aide_design/materials_database.py
@@ -10,11 +10,11 @@ import math
 import numpy as np
 
 try:
-    from AIDE.units import unit_registry as u
-    from AIDE import utility as ut
+    from aide_design.units import unit_registry as u
+    from aide_design import utility as ut
 except ModuleNotFoundError:
-    from AIDE.units import unit_registry as u
-    from AIDE import utility as ut
+    from aide_design.units import unit_registry as u
+    from aide_design import utility as ut
 
 gravity = 9.80665 * u.m/u.s**2
 """Define the gravitational constant, in m/s²."""
