+++ aide_design/unit_definitions.txt
