+++ aide_design/unit_process_design/prefab/cdc_prefab.py
@@ -8,11 +8,11 @@ Created on Mon Jun 26 11:55:37 2017
 
 
 try:
-    from AIDE.units import unit_registry as u
-    from AIDE import utility as ut
+    from aide_design.units import unit_registry as u
+    from aide_design import utility as ut
 except ModuleNotFoundError:
-    from AIDE.units import unit_registry as u
-    from AIDE import utility as ut
+    from aide_design.units import unit_registry as u
+    from aide_design import utility as ut
 
 g=9.80665*(u.m/(u.s**2))
 #==============================================================================
