+++ aide_design/unit_process_draw/cdc_prefab_sp17 v1.iges
