+++ setup.py
@@ -1,12 +1,12 @@
 from setuptools import setup
 
-setup(name='AIDE',
+setup(name='aide_design',
       version='0.0',
       description='AguaClara Infrastructure Design Engine',
       url='https://github.com/AguaClara/aguaclara_design',
       author='AguaClara at Cornell',
       author_email='aguaclara@cornell.edu',
       license='MIT',
-      packages=['AIDE'],
+      packages=['aide_design'],
       install_requires=['pint','numpy','pandas','matplotlib'],
       zip_safe=False)
